arxiv/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
pymupdf>=1.23.0
pdfplumber>=0.7.0
pillow>=8.0.0
pandas>=1.3.0
arxiv>=1.4.0
pypandoc>=1.6.0
pathlib2>=2.3.0